requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
click==8.3.1
fastapi==0.123.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dotenv==1.2.1
PyYAML==6.0.3
passlib[bcrypt]==1.7.4
bcrypt<4
python-jose==3.3.0
SQLAlchemy==2.0.44
alembic==1.13.2
asyncpg==0.30.0
sqlmodel==0.0.27
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
yt-dlp
aiofiles
Pillow==10.3.0

# Enhanced music APIs for better track metadata
spotipy==2.23.0          # Official Spotify Python library with better access
lyricsgenius==3.0.1      # Genius.com lyrics and additional track data
discogs-client==2.3.0      # Discogs database API for alternative metadata
beautifulsoup4==4.12.0   # For web scraping additional data
